+++ packages/ckbtc/src/index.ts
@@ -1,4 +1,12 @@
 export type {
+  block_hash,
+  block_height,
+  get_utxos_response,
+  outpoint,
+  satoshi,
+  utxo,
+} from "../candid/bitcoin";
+export type {
   Account,
   MinterInfo,
   PendingUtxo,
@@ -9,10 +17,12 @@ export type {
   UtxoStatus,
   Account as WithdrawalAccount,
 } from "../candid/minter";
+export { BitcoinCanister } from "./bitcoin.canister";
 export * from "./enums/btc.enums";
 export * from "./errors/btc.errors";
 export * from "./errors/minter.errors";
 export { CkBTCMinterCanister } from "./minter.canister";
+export * from "./types/bitcoin.params";
 export * from "./types/btc";
 export * from "./types/minter.params";
 export * from "./types/minter.responses";