+++ packages/ckbtc/src/types/canister.options.ts
@@ -1,8 +1,8 @@
 import type { Principal } from "@dfinity/principal";
 import type { CanisterOptions } from "@dfinity/utils";
 
-export interface CkBTCMinterCanisterOptions<T>
+export interface CkBTCCanisterOptions<T>
   extends Omit<CanisterOptions<T>, "canisterId"> {
-  // The canister's ID is mandatory to instantiate an ckBTC minter.
+  // The canister's ID is mandatory to instantiate a ckBTC related canister.
   canisterId: Principal;
 }